+++ activiti/src/main/java/com/maoding/activiti/dto/FlowTaskGroupDTO.java
@@ -13,10 +13,10 @@ import java.util.List;
  **/
 public class FlowTaskGroupDTO extends CoreShowDTO {
     /** 条件最小值 */
-    private Integer minValue;
+    private Double minValue;
 
     /** 条件最大值 */
-    private Integer maxValue;
+    private Double maxValue;
 
     /** 显示条件 */
     private String title;
@@ -32,19 +32,19 @@ public class FlowTaskGroupDTO extends CoreShowDTO {
         this.title = title;
     }
 
-    public Integer getMinValue() {
+    public Double getMinValue() {
         return minValue;
     }
 
-    public void setMinValue(Integer minValue) {
+    public void setMinValue(Double minValue) {
         this.minValue = minValue;
     }
 
-    public Integer getMaxValue() {
+    public Double getMaxValue() {
         return maxValue;
     }
 
-    public void setMaxValue(Integer maxValue) {
+    public void setMaxValue(Double maxValue) {
         this.maxValue = maxValue;
     }
